requirements.txt
discord.py>=2.5.0
python-dotenv>=1.0.1
openai>=1.51.0
pymongo>=4.10.0
PyNacl>=1.4.0
